+++ app/src/test/java/com/github/dach83/gasmetering/fake/FakeAbonentsRepository.kt
@@ -3,7 +3,7 @@ package com.github.dach83.gasmetering.fake
 import android.net.Uri
 import com.github.dach83.gasmetering.features.abonents.domain.model.Abonent
 import com.github.dach83.gasmetering.features.abonents.domain.repository.AbonentsRepository
-import com.github.dach83.gasmetering.models.fakeAbonents
+import com.github.dach83.sharedtest.models.fakeAbonents
 
 class FakeAbonentsRepository : AbonentsRepository {
 
@@ -12,7 +12,7 @@ class FakeAbonentsRepository : AbonentsRepository {
 
     override suspend fun loadAbonents(
         excelUri: Uri,
-        onLoading: suspend (progress: Byte, abonents: List<Abonent>) -> Unit
+        onLoading: suspend (progress: Int, abonents: List<Abonent>) -> Unit
     ) {
         error?.let { throw it }
         onLoading(100, abonents)
